Client API Refactor

+++ ozone-framework-python-server/stacks/views.py
@@ -3,6 +3,8 @@ from rest_framework.decorators import action
 from rest_framework.permissions import IsAuthenticated, IsAdminUser
 from rest_framework.response import Response
 from django_filters.rest_framework import DjangoFilterBackend
+
+from config.owf_mixins import mixins
 from .models import Stack, StackGroups
 from .serializers import StackSerializer, StackGroupsSerializer, StackGroupsSerializerList
 from dashboards.permissions import IsStackOwner
@@ -62,11 +64,12 @@ class StackAdminViewSet(viewsets.ModelViewSet):
         all_stack_dashboard_ids = list(all_stack_dashboards.values_list("id", flat=True))
         DomainMapping.objects.filter(src_id__in=all_stack_dashboard_ids, src_type=MappingType.dashboard).delete()
         DomainMapping.objects.filter(dest_id__in=all_stack_dashboard_ids, dest_type=MappingType.dashboard).delete()
-        stack.default_group.delete()
+        if stack.default_group:
+            stack.default_group.delete()
         stack.delete()
 
 
-class StackGroupsViewSet(viewsets.ModelViewSet):
+class StackGroupsViewSet(mixins.BulkDestroyModelMixin, viewsets.ModelViewSet):
     """
     API endpoint that allows stack groups to be viewed or edited
     """
@@ -80,3 +83,31 @@ class StackGroupsViewSet(viewsets.ModelViewSet):
             return StackGroupsSerializerList
         else:
             return StackGroupsSerializer
+
+    # Support deletion using stack_id and group_id or list of group_ids, so that we don't have to
+    # modify the frontend to keep track of the PK of the intermediary tables
+    def destroy(self, request, *args, **kwargs):
+        try:
+            filters = {'stack_id': request.data.get('stack_id')}
+            id_list = request.data and request.data.get('id')
+            group_id = request.data and request.data.get('group_id')
+            group_ids = request.data and request.data.get('group_ids')
+
+            if id_list:
+                filters.update({
+                    'id__in': list(map(int, id_list))
+                })
+            elif group_id:
+                filters.update({
+                    'group_id': group_id
+                })
+            elif group_ids:
+                filters.update({
+                    'group_id__in': list(map(int, group_ids))
+                })
+            else:
+                return Response({'detail': 'missing group_id or group_ids'}, status=status.HTTP_400_BAD_REQUEST)
+
+            return self.bulk_destroy(request, **filters)
+        except (ValueError,):
+            return Response(status=status.HTTP_400_BAD_REQUEST)
